fix: 대시보드 랜더링 이슈 수정
