+++ src/app/(app)/dashboard/[id]/components/DashboardView/styles.ts
